fixed grids with no screen distortion
Only part of the grating was shown when "screen distortion" was not selected and and an angle different than k*90 degrees (with k=0,1...) was selected. This is because the texture was not centered to the destination rectangle (the screen) when Screen('DrawTextures',...) was called. I used the command CenterRect to solve the issue.
Also, I fixed the angles and direction of the moving gratings: now, angles are measured as in the goniometric circle (counterclockwise to the x axis: 0 degrees is the angle with cos=1). So, when the angle is 0, gratings slide toward the positive direction of the x axis (to the right); when the angle is 90°, they move toward the positive direction of the y axis (upward), and so on...