major update: re-organization of slave
slave reorganized: added a wrapper function to call visual stimuli. This avoids the mess of parameters there was before. Now editing function is much easier, and you probably don't even have to touch any line of code in the "slave" app when you need to change the number of parameters of visual stimuli.